+++ src/main/java/hosh/runtime/ExternalCommand.java
@@ -37,7 +37,6 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Locale;
 import java.util.Map;
-import java.util.Optional;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 
@@ -113,12 +112,7 @@ public class ExternalCommand implements Command, StateAware {
 	private void pipeChannelToOutputStream(InputChannel in, OutputStream outputStream) {
 		Locale locale = Locale.getDefault();
 		try (PrintWriter pw = new PrintWriter(new OutputStreamWriter(outputStream, StandardCharsets.UTF_8), true)) {
-			while (true) {
-				Optional<Record> recv = in.recv();
-				if (recv.isEmpty()) {
-					break;
-				}
-				Record record = recv.get();
+			for (Record record : InputChannel.iterate(in)) {
 				record.print(pw, locale);
 				pw.println();
 			}