+++ src/main/java/hosh/runtime/NullChannel.java
@@ -42,4 +42,5 @@ public class NullChannel implements InputChannel {
 	public String toString() {
 		return "NullChannel[]";
 	}
+
 }